* поправил конфликты * сделал кнопку "выйти" не только красивой, но и рабочей
