+++ src/main/java/ru/vkr/vkr/controller/AdminController.java
@@ -2,11 +2,6 @@ package ru.vkr.vkr.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import ru.vkr.vkr.service.UserService;
 
 @Controller
