+++ src/main/java/ru/vkr/vkr/controller/TeacherController.java
@@ -1,14 +1,8 @@
 package ru.vkr.vkr.controller;
 
-import edu.csus.ecs.pc2.api.ITeam;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.MissingPathVariableException;
 import org.springframework.web.bind.annotation.GetMapping;
-import ru.vkr.vkr.contest.Test;
-
-import java.util.ArrayList;
-import java.util.List;
 
 @Controller
 public class TeacherController {
@@ -37,12 +31,8 @@ public class TeacherController {
         return "course";
     }
 
-
     @GetMapping("/teacher/theme")
     public String testTheme() {
         return "theme";
     }
-
-
-
 }
