+++ src/main/resources/templates/base.html
@@ -7,9 +7,12 @@
         <meta name = "viewport" content = "width=device-width, initial-scale=1.0">
         <title th:replace="${title}">Title</title>
 
+        <link rel="shortcut icon" th:href="@{/img/altstu-logo.png}"/>
+
         <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/css/bootstrap.min.css" integrity="sha384-Vkoo8x4CGsO3+Hhxv8T/Q5PaXtkKtu6ug5TOeNV6gBiFeWPGFN9MuhOf23Q9Ifjh" crossorigin="anonymous">
-        <link rel="stylesheet" th:href="@{/css/font-awesome.min.css}">
-        <link rel="stylesheet" th:href="@{/css/main.css}">
+        <link rel="stylesheet" th:href="@{/css/font-awesome.min.css}"/>
+        <link rel="stylesheet" th:href="@{/css/main.css}"/>
+        <link rel="stylesheet" th:href="@{/css/signin.css}"/>
     </head>
 
 
@@ -41,7 +44,9 @@
                         <!-- Отображается либо "Войти" либо Имя -->
                         <a href=""><i class="fa fa-user" aria-hidden="true"></i>&nbsp;Иванов Иван</a>
                         &nbsp;&nbsp;|&nbsp;&nbsp;
-                        <a href="" data-toggle="modal" data-target="#exampleModal"><i class="fa fa-sign-out" aria-hidden="true"></i>&nbsp;Выйти</a>
+                        <form id="form" th:action="@{/logout}" method="post">
+                            <a href="javascript:void(0)" id="submit" data-toggle="modal" data-target="#exampleModal"><i class="fa fa-sign-out" aria-hidden="true"></i>&nbsp;Выйти</a>
+                        </form>
                     </div>
                 </nav>
             </div>
